M5L1: fix tests.
