+++ ok-marketplace-biz/src/commonTest/kotlin/ru/otus/otuskotlin/marketplace/biz/validation/BizValidationCreateTest.kt
@@ -3,7 +3,6 @@ package ru.otus.otuskotlin.marketplace.biz.validation
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import ru.otus.otuskotlin.marketplace.backend.repository.inmemory.AdRepoStub
 import ru.otus.otuskotlin.marketplace.biz.MkplAdProcessor
-import ru.otus.otuskotlin.marketplace.biz.validation.validation.*
 import ru.otus.otuskotlin.marketplace.common.MkplCorSettings
 import ru.otus.otuskotlin.marketplace.common.models.MkplCommand
 import kotlin.test.Test
