+++ ok-marketplace-biz/src/commonTest/kotlin/ru/otus/otuskotlin/marketplace/biz/validation/BizValidationDeleteTest.kt
@@ -1,20 +1,28 @@
-package ru.otus.otuskotlin.marketplace.biz.validation.validation
+package ru.otus.otuskotlin.marketplace.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import ru.otus.otuskotlin.marketplace.backend.repository.inmemory.AdRepoStub
 import ru.otus.otuskotlin.marketplace.biz.MkplAdProcessor
+import ru.otus.otuskotlin.marketplace.common.MkplCorSettings
 import ru.otus.otuskotlin.marketplace.common.models.MkplCommand
 import kotlin.test.Test
 
 @OptIn(ExperimentalCoroutinesApi::class)
-class BizValidationReadTest {
+class BizValidationDeleteTest {
 
-    private val command = MkplCommand.READ
-    private val processor by lazy { MkplAdProcessor() }
+    private val command = MkplCommand.DELETE
+    private val settings by lazy {
+        MkplCorSettings(
+            repoTest = AdRepoStub()
+        )
+    }
+    private val processor by lazy { MkplAdProcessor(settings) }
 
     @Test fun correctId() = validationIdCorrect(command, processor)
     @Test fun trimId() = validationIdTrim(command, processor)
     @Test fun emptyId() = validationIdEmpty(command, processor)
     @Test fun badFormatId() = validationIdFormat(command, processor)
+
 
 }
