+++ ok-marketplace-biz/src/commonTest/kotlin/ru/otus/otuskotlin/marketplace/biz/validation/BizValidationReadTest.kt
@@ -1,21 +1,27 @@
-package ru.otus.otuskotlin.marketplace.biz.validation.validation
+package ru.otus.otuskotlin.marketplace.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import ru.otus.otuskotlin.marketplace.backend.repository.inmemory.AdRepoStub
 import ru.otus.otuskotlin.marketplace.biz.MkplAdProcessor
+import ru.otus.otuskotlin.marketplace.common.MkplCorSettings
 import ru.otus.otuskotlin.marketplace.common.models.MkplCommand
 import kotlin.test.Test
 
 @OptIn(ExperimentalCoroutinesApi::class)
-class BizValidationDeleteTest {
+class BizValidationReadTest {
 
-    private val command = MkplCommand.DELETE
-    private val processor by lazy { MkplAdProcessor() }
+    private val command = MkplCommand.READ
+    private val settings by lazy {
+        MkplCorSettings(
+            repoTest = AdRepoStub()
+        )
+    }
+    private val processor by lazy { MkplAdProcessor(settings) }
 
     @Test fun correctId() = validationIdCorrect(command, processor)
     @Test fun trimId() = validationIdTrim(command, processor)
     @Test fun emptyId() = validationIdEmpty(command, processor)
     @Test fun badFormatId() = validationIdFormat(command, processor)
-
 
 }
