+++ ok-marketplace-biz/src/commonTest/kotlin/ru/otus/otuskotlin/marketplace/biz/validation/BizValidationSearchTest.kt
@@ -1,9 +1,11 @@
-package ru.otus.otuskotlin.marketplace.biz.validation.validation
+package ru.otus.otuskotlin.marketplace.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
+import ru.otus.otuskotlin.marketplace.backend.repository.inmemory.AdRepoStub
 import ru.otus.otuskotlin.marketplace.biz.MkplAdProcessor
 import ru.otus.otuskotlin.marketplace.common.MkplContext
+import ru.otus.otuskotlin.marketplace.common.MkplCorSettings
 import ru.otus.otuskotlin.marketplace.common.models.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -13,7 +15,12 @@ import kotlin.test.assertNotEquals
 class BizValidationSearchTest {
 
     private val command = MkplCommand.SEARCH
-    private val processor by lazy { MkplAdProcessor() }
+    private val settings by lazy {
+        MkplCorSettings(
+            repoTest = AdRepoStub()
+        )
+    }
+    private val processor by lazy { MkplAdProcessor(settings) }
 
     @Test
     fun correctEmpty() = runTest {
