+++ ok-marketplace-biz/src/commonTest/kotlin/ru/otus/otuskotlin/marketplace/biz/validation/BizValidationUpdateTest.kt
@@ -1,7 +1,9 @@
-package ru.otus.otuskotlin.marketplace.biz.validation.validation
+package ru.otus.otuskotlin.marketplace.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import ru.otus.otuskotlin.marketplace.backend.repository.inmemory.AdRepoStub
 import ru.otus.otuskotlin.marketplace.biz.MkplAdProcessor
+import ru.otus.otuskotlin.marketplace.common.MkplCorSettings
 import ru.otus.otuskotlin.marketplace.common.models.MkplCommand
 import kotlin.test.Test
 
@@ -9,7 +11,12 @@ import kotlin.test.Test
 class BizValidationUpdateTest {
 
     private val command = MkplCommand.UPDATE
-    private val processor by lazy { MkplAdProcessor() }
+    private val settings by lazy {
+        MkplCorSettings(
+            repoTest = AdRepoStub()
+        )
+    }
+    private val processor by lazy { MkplAdProcessor(settings) }
 
     @Test fun correctTitle() = validationTitleCorrect(command, processor)
     @Test fun trimTitle() = validationTitleTrim(command, processor)
