+++ ok-marketplace-biz/src/commonTest/kotlin/ru/otus/otuskotlin/marketplace/biz/validation/ValidationBadDescription.kt
@@ -1,22 +1,23 @@
-package ru.otus.otuskotlin.marketplace.biz.validation.validation
+package ru.otus.otuskotlin.marketplace.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import ru.otus.otuskotlin.marketplace.biz.MkplAdProcessor
 import ru.otus.otuskotlin.marketplace.common.MkplContext
 import ru.otus.otuskotlin.marketplace.common.models.*
+import ru.otus.otuskotlin.marketplace.stubs.MkplAdStub
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
-
+private val stub = MkplAdStub.get()
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationIdCorrect(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationDescriptionCorrect(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = MkplAdId("123-234-abc-ABC"),
+            id = stub.id,
             title = "abc",
             description = "abc",
             adType = MkplDealSide.DEMAND,
@@ -26,18 +27,19 @@ fun validationIdCorrect(command: MkplCommand, processor: MkplAdProcessor) = runT
     processor.exec(ctx)
     assertEquals(0, ctx.errors.size)
     assertNotEquals(MkplState.FAILING, ctx.state)
+    assertEquals("abc", ctx.adValidated.description)
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationIdTrim(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationDescriptionTrim(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = MkplAdId(" \n\t 123-234-abc-ABC \n\t "),
+            id = stub.id,
             title = "abc",
-            description = "abc",
+            description = " \n\tabc \n\t",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
         ),
@@ -45,18 +47,19 @@ fun validationIdTrim(command: MkplCommand, processor: MkplAdProcessor) = runTest
     processor.exec(ctx)
     assertEquals(0, ctx.errors.size)
     assertNotEquals(MkplState.FAILING, ctx.state)
+    assertEquals("abc", ctx.adValidated.description)
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationIdEmpty(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationDescriptionEmpty(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = MkplAdId(""),
+            id = stub.id,
             title = "abc",
-            description = "abc",
+            description = "",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
         ),
@@ -65,20 +68,20 @@ fun validationIdEmpty(command: MkplCommand, processor: MkplAdProcessor) = runTes
     assertEquals(1, ctx.errors.size)
     assertEquals(MkplState.FAILING, ctx.state)
     val error = ctx.errors.firstOrNull()
-    assertEquals("id", error?.field)
-    assertContains(error?.message ?: "", "id")
+    assertEquals("description", error?.field)
+    assertContains(error?.message ?: "", "description")
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationIdFormat(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationDescriptionSymbols(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = MkplAdId("!@#\$%^&*(),.{}"),
+            id = stub.id,
             title = "abc",
-            description = "abc",
+            description = "!@#$%^&*(),.{}",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
         ),
@@ -87,6 +90,6 @@ fun validationIdFormat(command: MkplCommand, processor: MkplAdProcessor) = runTe
     assertEquals(1, ctx.errors.size)
     assertEquals(MkplState.FAILING, ctx.state)
     val error = ctx.errors.firstOrNull()
-    assertEquals("id", error?.field)
-    assertContains(error?.message ?: "", "id")
+    assertEquals("description", error?.field)
+    assertContains(error?.message ?: "", "description")
 }
