+++ ok-marketplace-biz/src/commonTest/kotlin/ru/otus/otuskotlin/marketplace/biz/validation/ValidationBadId.kt
@@ -1,25 +1,22 @@
-package ru.otus.otuskotlin.marketplace.biz.validation.validation
+package ru.otus.otuskotlin.marketplace.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import ru.otus.otuskotlin.marketplace.biz.MkplAdProcessor
 import ru.otus.otuskotlin.marketplace.common.MkplContext
 import ru.otus.otuskotlin.marketplace.common.models.*
-import ru.otus.otuskotlin.marketplace.stubs.MkplAdStub
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
 
-private val stub = MkplAdStub.get()
-
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationTitleCorrect(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationIdCorrect(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = stub.id,
+            id = MkplAdId("123-234-abc-ABC"),
             title = "abc",
             description = "abc",
             adType = MkplDealSide.DEMAND,
@@ -29,18 +26,17 @@ fun validationTitleCorrect(command: MkplCommand, processor: MkplAdProcessor) = r
     processor.exec(ctx)
     assertEquals(0, ctx.errors.size)
     assertNotEquals(MkplState.FAILING, ctx.state)
-    assertEquals("abc", ctx.adValidated.title)
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationTitleTrim(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationIdTrim(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = stub.id,
-            title = " \n\t abc \t\n ",
+            id = MkplAdId(" \n\t 123-234-abc-ABC \n\t "),
+            title = "abc",
             description = "abc",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
@@ -49,18 +45,17 @@ fun validationTitleTrim(command: MkplCommand, processor: MkplAdProcessor) = runT
     processor.exec(ctx)
     assertEquals(0, ctx.errors.size)
     assertNotEquals(MkplState.FAILING, ctx.state)
-    assertEquals("abc", ctx.adValidated.title)
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationTitleEmpty(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationIdEmpty(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = stub.id,
-            title = "",
+            id = MkplAdId(""),
+            title = "abc",
             description = "abc",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
@@ -70,19 +65,19 @@ fun validationTitleEmpty(command: MkplCommand, processor: MkplAdProcessor) = run
     assertEquals(1, ctx.errors.size)
     assertEquals(MkplState.FAILING, ctx.state)
     val error = ctx.errors.firstOrNull()
-    assertEquals("title", error?.field)
-    assertContains(error?.message ?: "", "title")
+    assertEquals("id", error?.field)
+    assertContains(error?.message ?: "", "id")
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationTitleSymbols(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationIdFormat(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = MkplAdId("123"),
-            title = "!@#$%^&*(),.{}",
+            id = MkplAdId("!@#\$%^&*(),.{}"),
+            title = "abc",
             description = "abc",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
@@ -92,6 +87,6 @@ fun validationTitleSymbols(command: MkplCommand, processor: MkplAdProcessor) = r
     assertEquals(1, ctx.errors.size)
     assertEquals(MkplState.FAILING, ctx.state)
     val error = ctx.errors.firstOrNull()
-    assertEquals("title", error?.field)
-    assertContains(error?.message ?: "", "title")
+    assertEquals("id", error?.field)
+    assertContains(error?.message ?: "", "id")
 }
