+++ ok-marketplace-biz/src/commonTest/kotlin/ru/otus/otuskotlin/marketplace/biz/validation/ValidationBadTitle.kt
@@ -1,4 +1,4 @@
-package ru.otus.otuskotlin.marketplace.biz.validation.validation
+package ru.otus.otuskotlin.marketplace.biz.validation
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
@@ -13,7 +13,7 @@ import kotlin.test.assertNotEquals
 private val stub = MkplAdStub.get()
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationDescriptionCorrect(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationTitleCorrect(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
@@ -29,19 +29,19 @@ fun validationDescriptionCorrect(command: MkplCommand, processor: MkplAdProcesso
     processor.exec(ctx)
     assertEquals(0, ctx.errors.size)
     assertNotEquals(MkplState.FAILING, ctx.state)
-    assertEquals("abc", ctx.adValidated.description)
+    assertEquals("abc", ctx.adValidated.title)
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationDescriptionTrim(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationTitleTrim(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
             id = stub.id,
-            title = "abc",
-            description = " \n\tabc \n\t",
+            title = " \n\t abc \t\n ",
+            description = "abc",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
         ),
@@ -49,19 +49,19 @@ fun validationDescriptionTrim(command: MkplCommand, processor: MkplAdProcessor) 
     processor.exec(ctx)
     assertEquals(0, ctx.errors.size)
     assertNotEquals(MkplState.FAILING, ctx.state)
-    assertEquals("abc", ctx.adValidated.description)
+    assertEquals("abc", ctx.adValidated.title)
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationDescriptionEmpty(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationTitleEmpty(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
             id = stub.id,
-            title = "abc",
-            description = "",
+            title = "",
+            description = "abc",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
         ),
@@ -70,20 +70,20 @@ fun validationDescriptionEmpty(command: MkplCommand, processor: MkplAdProcessor)
     assertEquals(1, ctx.errors.size)
     assertEquals(MkplState.FAILING, ctx.state)
     val error = ctx.errors.firstOrNull()
-    assertEquals("description", error?.field)
-    assertContains(error?.message ?: "", "description")
+    assertEquals("title", error?.field)
+    assertContains(error?.message ?: "", "title")
 }
 
 @OptIn(ExperimentalCoroutinesApi::class)
-fun validationDescriptionSymbols(command: MkplCommand, processor: MkplAdProcessor) = runTest {
+fun validationTitleSymbols(command: MkplCommand, processor: MkplAdProcessor) = runTest {
     val ctx = MkplContext(
         command = command,
         state = MkplState.NONE,
         workMode = MkplWorkMode.TEST,
         adRequest = MkplAd(
-            id = stub.id,
-            title = "abc",
-            description = "!@#$%^&*(),.{}",
+            id = MkplAdId("123"),
+            title = "!@#$%^&*(),.{}",
+            description = "abc",
             adType = MkplDealSide.DEMAND,
             visibility = MkplVisibility.VISIBLE_PUBLIC,
         ),
@@ -92,6 +92,6 @@ fun validationDescriptionSymbols(command: MkplCommand, processor: MkplAdProcesso
     assertEquals(1, ctx.errors.size)
     assertEquals(MkplState.FAILING, ctx.state)
     val error = ctx.errors.firstOrNull()
-    assertEquals("description", error?.field)
-    assertContains(error?.message ?: "", "description")
+    assertEquals("title", error?.field)
+    assertContains(error?.message ?: "", "title")
 }
